feat: Add getSales endpoint to saleRoutes.js
This commit adds a new endpoint `getSales` to the `saleRoutes.js` file in the `api/routes` directory. The `getSales` endpoint retrieves all sales data from the database and populates the `product` field using the `populate` method. This allows the response to include the details of the associated product for each sale.